Usec2Sysclock fixes

+++ Core/HLE/sceKernelTime.h
@@ -25,7 +25,7 @@ u32 sceKernelGetSystemTimeLow();
 u64 sceKernelGetSystemTimeWide();
 int sceKernelSysClock2USec(u32 sysclockPtr, u32 highPtr, u32 lowPtr);
 int sceKernelSysClock2USecWide(u32 lowClock, u32 highClock, u32 lowPtr, u32 highPtr);
-u32 sceKernelUSec2SysClockWide(u32 usec);
+u64 sceKernelUSec2SysClockWide(u32 usec);
 u32 sceKernelLibcClock();
 
 void __KernelTimeInit();
